#RI-5033 - Refactor key details version 2
