+++ redisinsight/ui/src/pages/browser/modules/index.ts
@@ -1,2 +1,3 @@
 export { KeyDetails } from './key-details'
 export { KeyDetailsHeader } from './key-details-header'
+export type { KeyDetailsHeaderProps } from './key-details-header'
