+++ redisinsight/ui/src/pages/browser/modules/key-details-header/index.ts
@@ -1,1 +1,2 @@
 export { KeyDetailsHeader } from './KeyDetailsHeader'
+export type { KeyDetailsHeaderProps } from './KeyDetailsHeader'
