+++ redisinsight/ui/src/pages/browser/modules/key-details/KeyDetails.tsx
@@ -1,4 +1,4 @@
-import React, { useEffect, useState } from 'react'
+import React, { useEffect } from 'react'
 import { isNull, isUndefined } from 'lodash'
 import { useDispatch, useSelector } from 'react-redux'
 import { useParams } from 'react-router-dom'
@@ -10,17 +10,13 @@ import {
   selectedKeyDataSelector,
   selectedKeySelector,
 } from 'uiSrc/slices/browser/keys'
-import { KeyTypes, STREAM_ADD_GROUP_VIEW_TYPES } from 'uiSrc/constants'
+import { KeyTypes } from 'uiSrc/constants'
 
 import { getBasedOnViewTypeEvent, sendEventTelemetry, TelemetryEvent } from 'uiSrc/telemetry'
 import { RedisResponseBuffer } from 'uiSrc/slices/interfaces'
 import { Nullable } from 'uiSrc/utils'
-import { KeyDetailsHeader } from 'uiSrc/pages/browser/modules'
-import { streamSelector } from 'uiSrc/slices/browser/stream'
 import { NoKeySelected } from './components/no-key-selected'
 import { DynamicTypeDetails } from './components/dynamic-type-details'
-import { AddItemsPanel } from './components/add-items-panel'
-import { RemoveListElements } from './components/key-details-remove-items'
 
 import styles from './styles.module.scss'
 
@@ -39,8 +35,6 @@ export interface Props {
 const KeyDetails = (props: Props) => {
   const {
     onCloseKey,
-    onEditKey,
-    onRemoveKey,
     keyProp,
     totalKeys,
     keysLastRefreshTime,
@@ -50,14 +44,9 @@ const KeyDetails = (props: Props) => {
   const { viewType } = useSelector(keysSelector)
   const { loading, error = '', data } = useSelector(selectedKeySelector)
   const isKeySelected = !isNull(useSelector(selectedKeyDataSelector))
-  const { viewType: streamViewType } = useSelector(streamSelector)
   const { type: keyType, name: keyName, length: keyLength } = useSelector(selectedKeyDataSelector) ?? {
     type: KeyTypes.String,
   }
-
-  const [isAddItemPanelOpen, setIsAddItemPanelOpen] = useState<boolean>(false)
-  const [isRemoveItemPanelOpen, setIsRemoveItemPanelOpen] = useState<boolean>(false)
-  const [editItem, setEditItem] = useState<boolean>(false)
 
   const dispatch = useDispatch()
 
@@ -87,48 +76,32 @@ const KeyDetails = (props: Props) => {
     }
   }, [keyName])
 
-  const openAddItemPanel = () => {
-    setIsRemoveItemPanelOpen(false)
-    setIsAddItemPanelOpen(true)
-    if (!STREAM_ADD_GROUP_VIEW_TYPES.includes(streamViewType)) {
-      sendEventTelemetry({
-        event: getBasedOnViewTypeEvent(
-          viewType,
-          TelemetryEvent.BROWSER_KEY_ADD_VALUE_CLICKED,
-          TelemetryEvent.TREE_VIEW_KEY_ADD_VALUE_CLICKED
-        ),
-        eventData: {
-          databaseId: instanceId,
-          keyType
-        }
-      })
-    }
+  const onCloseAddItemPanel = () => {
+    sendEventTelemetry({
+      event: getBasedOnViewTypeEvent(
+        viewType,
+        TelemetryEvent.BROWSER_KEY_ADD_VALUE_CANCELLED,
+        TelemetryEvent.TREE_VIEW_KEY_ADD_VALUE_CANCELLED,
+      ),
+      eventData: {
+        databaseId: instanceId,
+        keyType,
+      }
+    })
   }
 
-  const openRemoveItemPanel = () => {
-    setIsAddItemPanelOpen(false)
-    setIsRemoveItemPanelOpen(true)
-  }
-
-  const closeAddItemPanel = (isCancelled?: boolean) => {
-    if (isCancelled && isAddItemPanelOpen && !STREAM_ADD_GROUP_VIEW_TYPES.includes(streamViewType)) {
-      sendEventTelemetry({
-        event: getBasedOnViewTypeEvent(
-          viewType,
-          TelemetryEvent.BROWSER_KEY_ADD_VALUE_CANCELLED,
-          TelemetryEvent.TREE_VIEW_KEY_ADD_VALUE_CANCELLED,
-        ),
-        eventData: {
-          databaseId: instanceId,
-          keyType,
-        }
-      })
-    }
-    setIsAddItemPanelOpen(false)
-  }
-
-  const closeRemoveItemPanel = () => {
-    setIsRemoveItemPanelOpen(false)
+  const onOpenAddItemPanel = () => {
+    sendEventTelemetry({
+      event: getBasedOnViewTypeEvent(
+        viewType,
+        TelemetryEvent.BROWSER_KEY_ADD_VALUE_CLICKED,
+        TelemetryEvent.TREE_VIEW_KEY_ADD_VALUE_CLICKED
+      ),
+      eventData: {
+        databaseId: instanceId,
+        keyType,
+      }
+    })
   }
 
   return (
@@ -143,47 +116,12 @@ const KeyDetails = (props: Props) => {
             onClosePanel={onCloseKey}
           />
         ) : (
-          <div className="fluid flex-column relative">
-            <KeyDetailsHeader
-              {...props}
-              key="key-details-header"
-              keyType={keyType}
-              onAddItem={openAddItemPanel}
-              onRemoveItem={openRemoveItemPanel}
-              onEditItem={() => setEditItem(!editItem)}
-              onRemoveKey={onRemoveKey}
-              onClose={onCloseKey}
-              onEditKey={onEditKey}
-            />
-            <div className="key-details-body" key="key-details-body">
-              {!loading && (
-                <div className="flex-column" style={{ flex: '1', height: '100%' }}>
-                  <DynamicTypeDetails
-                    selectedKeyType={keyType}
-                    isAddItemPanelOpen={isAddItemPanelOpen}
-                    onRemoveKey={onRemoveKey}
-                    editItem={editItem}
-                    setEditItem={setEditItem}
-                    isRemoveItemPanelOpen={isRemoveItemPanelOpen}
-                  />
-                </div>
-              )}
-              {isAddItemPanelOpen && (
-                <AddItemsPanel
-                  selectedKeyType={keyType}
-                  streamViewType={streamViewType}
-                  closeAddItemPanel={closeAddItemPanel}
-                />
-              )}
-              {isRemoveItemPanelOpen && (
-                <div className={cx('formFooterBar', styles.contentActive)}>
-                  {keyType === KeyTypes.List && (
-                    <RemoveListElements onCancel={closeRemoveItemPanel} onRemoveKey={onRemoveKey} />
-                  )}
-                </div>
-              )}
-            </div>
-          </div>
+          <DynamicTypeDetails
+            {...props}
+            keyType={keyType}
+            onOpenAddItemPanel={onOpenAddItemPanel}
+            onCloseAddItemPanel={onCloseAddItemPanel}
+          />
         )}
       </div>
     </div>
