+++ redisinsight/ui/src/pages/browser/modules/key-details/components/dynamic-type-details/DynamicTypeDetails.spec.tsx
@@ -27,7 +27,7 @@ describe('DynamicTypeDetails', () => {
     (type: KeyTypes, testId: string) => {
       const { queryByTestId } = render(<DynamicTypeDetails
         {...instance(mockedProps)}
-        selectedKeyType={type}
+        keyType={type}
       />)
       expect(queryByTestId(testId)).toBeInTheDocument()
     })
