+++ redisinsight/ui/src/pages/browser/modules/key-details/components/dynamic-type-details/DynamicTypeDetails.tsx
@@ -1,58 +1,32 @@
 import React from 'react'
-import { useDispatch } from 'react-redux'
 import { KeyTypes, MODULES_KEY_TYPES_NAMES, ModulesKeyTypes } from 'uiSrc/constants'
-import { RedisResponseBuffer } from 'uiSrc/slices/interfaces'
-import { IFetchKeyArgs } from 'uiSrc/constants/prop-types/keys'
-import { refreshKey } from 'uiSrc/slices/browser/keys'
-import ZSetDetails from '../zset-details/ZSetDetails'
-import SetDetails from '../set-details/SetDetails'
-import StringDetails from '../string-details/StringDetails'
-import HashDetails from '../hash-details/HashDetails'
-import ListDetails from '../list-details/ListDetails'
-import RejsonDetailsWrapper from '../rejson-details/RejsonDetailsWrapper'
-import StreamDetailsWrapper from '../stream-details'
+import { KeyDetailsHeaderProps } from 'uiSrc/pages/browser/modules'
 import ModulesTypeDetails from '../modules-type-details/ModulesTypeDetails'
 import UnsupportedTypeDetails from '../unsupported-type-details/UnsupportedTypeDetails'
+import { RejsonDetailsWrapper } from '../rejson-details'
+import { StringDetails } from '../string-details'
+import { ZSetDetails } from '../zset-details'
+import { SetDetails } from '../set-details'
+import { HashDetails } from '../hash-details'
+import { ListDetails } from '../list-details'
+import { StreamDetails } from '../stream-details'
 
-export interface Props {
-  selectedKeyType: KeyTypes | ModulesKeyTypes
-  isAddItemPanelOpen: boolean
-  isRemoveItemPanelOpen: boolean
-  editItem: boolean
-  onRemoveKey: ()=> void
-  setEditItem: (isEdit: boolean) => void
+export interface Props extends KeyDetailsHeaderProps {
+  onOpenAddItemPanel: () => void
+  onCloseAddItemPanel: () => void
 }
 
 const DynamicTypeDetails = (props: Props) => {
-  const {
-    selectedKeyType,
-    isAddItemPanelOpen,
-    onRemoveKey,
-    isRemoveItemPanelOpen,
-    editItem,
-    setEditItem,
-  } = props
-
-  const dispatch = useDispatch()
-
-  const handleRefreshKey = (key: RedisResponseBuffer, type: KeyTypes | ModulesKeyTypes, args: IFetchKeyArgs) => {
-    dispatch(refreshKey(key, type, args))
-  }
+  const { keyType: selectedKeyType } = props
 
   const TypeDetails: any = {
-    [KeyTypes.ZSet]: <ZSetDetails isFooterOpen={isAddItemPanelOpen} onRemoveKey={onRemoveKey} />,
-    [KeyTypes.Set]: <SetDetails isFooterOpen={isAddItemPanelOpen} onRemoveKey={onRemoveKey} />,
-    [KeyTypes.String]: (
-      <StringDetails
-        isEditItem={editItem}
-        setIsEdit={(isEdit) => setEditItem(isEdit)}
-        onRefresh={handleRefreshKey}
-      />
-    ),
-    [KeyTypes.Hash]: <HashDetails isFooterOpen={isAddItemPanelOpen} onRemoveKey={onRemoveKey} />,
-    [KeyTypes.List]: <ListDetails isFooterOpen={isAddItemPanelOpen || isRemoveItemPanelOpen} />,
-    [KeyTypes.ReJSON]: <RejsonDetailsWrapper />,
-    [KeyTypes.Stream]: <StreamDetailsWrapper isFooterOpen={isAddItemPanelOpen} />,
+    [KeyTypes.ZSet]: <ZSetDetails {...props} />,
+    [KeyTypes.Set]: <SetDetails {...props} />,
+    [KeyTypes.String]: <StringDetails {...props} />,
+    [KeyTypes.Hash]: <HashDetails {...props} />,
+    [KeyTypes.List]: <ListDetails {...props} />,
+    [KeyTypes.ReJSON]: <RejsonDetailsWrapper {...props} />,
+    [KeyTypes.Stream]: <StreamDetails {...props} />,
   }
 
   // Supported key type
