+++ redisinsight/ui/src/pages/browser/modules/key-details/components/hash-details/hash-details-table/index.ts
@@ -1,0 +1,1 @@
+export { HashDetailsTable } from './HashDetailsTable'
