+++ redisinsight/ui/src/pages/browser/modules/key-details/components/hash-details/index.ts
@@ -1,0 +1,1 @@
+export { HashDetails } from './HashDetails'
