+++ redisinsight/ui/src/pages/browser/modules/key-details/components/key-details-add-items/add-list-elements/AddListElements.spec.tsx
@@ -1,6 +1,6 @@
 import React from 'react'
-import { fireEvent, render, screen } from 'uiSrc/utils/test-utils'
 import { instance, mock } from 'ts-mockito'
+import { fireEvent, render, screen } from 'uiSrc/utils/test-utils'
 import AddListElements, { HEAD_DESTINATION, Props } from './AddListElements'
 
 const mockedProps = mock<Props>()
