+++ redisinsight/ui/src/pages/browser/modules/key-details/components/list-details/index.ts
@@ -1,0 +1,1 @@
+export { ListDetails } from './ListDetails'
