+++ redisinsight/ui/src/pages/browser/modules/key-details/components/list-details/list-details-table/index.ts
@@ -1,0 +1,1 @@
+export { ListDetailsTable } from './ListDetailsTable'
