+++ redisinsight/ui/src/pages/browser/modules/key-details/components/list-details/styles.module.scss
@@ -1,41 +1,4 @@
-.container {
-  display: flex;
-  flex: 1;
-  width: 100%;
-  padding: 16px;
-  background-color: var(--euiColorEmptyShade);
-
-  .textAreaControls {
-    right: -56px !important;
-    bottom: -4px !important;
-    top: auto !important;
-    background-color: var(--euiPageBackgroundColor) !important;
-  }
-
-  .textArea {
-    background-color: var(--euiPageBackgroundColor) !important;
-    border-color: var(--euiColorPrimary) !important;
-    z-index: 3;
-    padding-left: 20px;
-    padding-bottom: 36px !important;
-    margin: -8px -6px -8px -20px !important;
-    min-width: calc(100% + 106px) !important;
-    font: normal normal normal 13px/18px Graphik, sans-serif;
-    min-height: 43px;
-    overflow: hidden;
-    overflow-wrap: break-word;
-    resize: none;
-
-    &:focus {
-      background-image: none !important;
-    }
-
-    &.areaWarning {
-      border-color: var(--euiColorWarningLight) !important;
-    }
-  }
-}
-
-.inlineItemEditor {
-  max-width: calc(100% - 20px);
+.contentActive {
+  border-color: var(--euiColorPrimary) !important;
+  border-bottom-width: 1px !important;
 }
