+++ redisinsight/ui/src/pages/browser/modules/key-details/components/rejson-details/index.ts
@@ -1,0 +1,1 @@
+export { RejsonDetailsWrapper } from './RejsonDetailsWrapper'
