+++ redisinsight/ui/src/pages/browser/modules/key-details/components/set-details/index.ts
@@ -1,0 +1,1 @@
+export { SetDetails } from './SetDetails'
