+++ redisinsight/ui/src/pages/browser/modules/key-details/components/set-details/set-details-table/index.ts
@@ -1,0 +1,1 @@
+export { SetDetailsTable } from './SetDetailsTable'
