+++ redisinsight/ui/src/pages/browser/modules/key-details/components/set-details/set-details-table/styles.module.scss
