+++ redisinsight/ui/src/pages/browser/modules/key-details/components/stream-details/index.ts
@@ -1,3 +1,1 @@
-import StreamDetailsWrapper from './StreamDetailsWrapper'
-
-export default StreamDetailsWrapper
+export { StreamDetails } from './StreamDetails'
