+++ redisinsight/ui/src/pages/browser/modules/key-details/components/stream-details/stream-details-body/index.ts
@@ -1,0 +1,1 @@
+export { StreamDetailsBody } from './StreamDetailsBody'
