+++ redisinsight/ui/src/pages/browser/modules/key-details/components/string-details/index.ts
@@ -1,0 +1,1 @@
+export { StringDetails } from './StringDetails'
