+++ redisinsight/ui/src/pages/browser/modules/key-details/components/string-details/string-details-table/index.ts
@@ -1,0 +1,1 @@
+export { StringDetailsTable } from './StringDetailsTable'
