+++ redisinsight/ui/src/pages/browser/modules/key-details/components/string-details/string-details-table/styles.module.scss
