+++ redisinsight/ui/src/pages/browser/modules/key-details/components/zset-details/index.ts
@@ -1,0 +1,1 @@
+export { ZSetDetails } from './ZSetDetails'
