+++ redisinsight/ui/src/pages/browser/modules/key-details/components/zset-details/zset-details-table/index.ts
@@ -1,0 +1,1 @@
+export { ZSetDetailsTable } from './ZSetDetailsTable'
